use pytest hook to decorate
